Add travis.yml (#34)
Adds Travis-CI configuration (issue #33)

Comments by sebix: "No fancy things like coverage and pep8 yet, but it works :) (for the first we cannot use the nosetests executable, as the c-code is not available then and we need to add this to the test_suite. For the last lot's of fixes are needed beforehand). See a successful build here: https://travis-ci.org/sebix/pyasn/builds/159705563"

Signed-off-by: Sebastian Wagner <sebix@sebix.at>